remove separate workflow file
