+++ pyproject.toml
@@ -44,7 +44,7 @@ test = [
 
 [project.scripts]
 # sets off the whole workflow in cpg-flow orchestrated Stages
-run_workflow = 'clinvarbitration.run_workflow:cli_main'
+run_workflow = 'clinvarbitration.stages:cli_main'
 
 # Step 1; re-summarise ClinVar using altered conflict resolution
 resummary = 'clinvarbitration.scripts.resummarise_clinvar:cli_main'
